# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API For Master Operasional Daerah Services",
        "title": "Operasional Daerah Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/",
    "paths": {
        "/opd/create": {
            "post": {
                "description": "Create a new opd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opd"
                ],
                "summary": "Create Opd",
                "parameters": [
                    {
                        "description": "Create Opd",
                        "name": "opd",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.OpdCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/opd/delete/{id}": {
            "delete": {
                "description": "Delete an existing opd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opd"
                ],
                "summary": "Delete Opd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opd ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/opd/detail/{id}": {
            "get": {
                "description": "Find By Id an existing opd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opd"
                ],
                "summary": "FindById opd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opd ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/opd/findall": {
            "get": {
                "description": "FindAll an existing opd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opd"
                ],
                "summary": "FindAll opd",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/opd/findallopd": {
            "get": {
                "description": "Find all only opd for dropdown",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opd"
                ],
                "summary": "Find all only opd",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/opd/findbyopd/{kode_opd}": {
            "get": {
                "description": "FindByKodeOpd an existing opd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opd"
                ],
                "summary": "FindByKodeOpd opd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kode OPD",
                        "name": "kode_opd",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/opd/update/{id}": {
            "put": {
                "description": "Update an existing opd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opd"
                ],
                "summary": "Update Opd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opd ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Opd",
                        "name": "opd",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.OpdUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "web.OpdCreateRequest": {
            "type": "object",
            "required": [
                "alamat",
                "kode_opd",
                "nama_kepala_opd",
                "nama_opd",
                "nip_kepala_opd",
                "singkatan"
            ],
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fax": {
                    "type": "string"
                },
                "kode_opd": {
                    "type": "string"
                },
                "nama_kepala_opd": {
                    "type": "string"
                },
                "nama_opd": {
                    "type": "string"
                },
                "nip_kepala_opd": {
                    "type": "string"
                },
                "pangkat_kepala": {
                    "type": "string"
                },
                "singkatan": {
                    "type": "string"
                },
                "telepon": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "web.OpdUpdateRequest": {
            "description": "Update Request OPD",
            "type": "object",
            "required": [
                "alamat",
                "kode_opd",
                "nama_kepala_opd",
                "nama_opd",
                "nip_kepala_opd",
                "singkatan"
            ],
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fax": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kode_opd": {
                    "type": "string"
                },
                "nama_kepala_opd": {
                    "type": "string"
                },
                "nama_opd": {
                    "type": "string"
                },
                "nip_kepala_opd": {
                    "type": "string"
                },
                "pangkat_kepala": {
                    "type": "string"
                },
                "singkatan": {
                    "type": "string"
                },
                "telepon": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "web.WebResponse": {
            "description": "Web Response",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}